VKIDSDK-507: Fix OneTap corner radius

+++ onetap-common/src/main/kotlin/com/vk/id/onetap/common/button/style/OneTapButtonCornersStyle.kt
@@ -1,17 +1,17 @@
 package com.vk.id.onetap.common.button.style
 
 public sealed class OneTapButtonCornersStyle(
-    public val radiusDp: Int
+    public val radiusDp: Float
 ) {
     public object Default : OneTapButtonCornersStyle(ROUNDED_RADIUS_DP)
     public object None : OneTapButtonCornersStyle(NONE_RADIUS_DP)
     public object Rounded : OneTapButtonCornersStyle(ROUNDED_RADIUS_DP)
     public object Round : OneTapButtonCornersStyle(ROUND_RADIUS_DP)
-    public class Custom(radiusDp: Int) : OneTapButtonCornersStyle(radiusDp)
+    public class Custom(radiusDp: Float) : OneTapButtonCornersStyle(radiusDp)
 
     private companion object {
-        private const val ROUNDED_RADIUS_DP = 8
-        private const val NONE_RADIUS_DP = 0
-        private const val ROUND_RADIUS_DP = 1000
+        private const val ROUNDED_RADIUS_DP = 8f
+        private const val NONE_RADIUS_DP = 0f
+        private const val ROUND_RADIUS_DP = 1000f
     }
 }
